+++ src/coalesce-vue-vuetify3/src/components/admin/c-admin-method.spec.tsx
@@ -1,11 +1,19 @@
 import { CAdminMethod } from "..";
+import $metadata from "@test-targets/metadata.g";
 import {
   PersonViewModel,
   PersonListViewModel,
   WeatherServiceViewModel,
 } from "@test-targets/viewmodels.g";
+import { defineComponent, h, nextTick } from "vue";
+import { mockEndpoint, mountWithCoalesceOptions } from "@test/util";
 
 describe("CAdminMethod", () => {
+  mockEndpoint("/Person/rename", () => ({
+    wasSuccessful: true,
+    object: { personId: 1, firstName: "A", lastName: "B" },
+  }));
+
   test("types", () => {
     const vm = new PersonViewModel();
     const listVm = new PersonListViewModel();
@@ -35,5 +43,45 @@ describe("CAdminMethod", () => {
     () => <CAdminMethod model={serviceVm} for="invalid" />;
     //@ts-expect-error method from wrong type
     () => <CAdminMethod model={serviceVm} for={rename} />;
+  });
+
+  test("renders registered admin input/display overrides for params and return", async () => {
+    const vm = new PersonViewModel();
+
+    const CustomInput = defineComponent({
+      name: "CustomMethodInput",
+      setup(_, { slots }) {
+        return () =>
+          h("div", { class: "custom-method-input" }, slots.default?.());
+      },
+    });
+
+    const CustomDisplay = defineComponent({
+      name: "CustomMethodDisplay",
+      setup() {
+        return () => h("div", { class: "custom-method-display" }, "custom");
+      },
+    });
+
+    const wrapper = mountWithCoalesceOptions(
+      () => <CAdminMethod model={vm} for="rename" />,
+      undefined,
+      {
+        adminOverrides: {
+          input: new Map([
+            [$metadata.types.Person.methods.rename.params.name, CustomInput],
+          ]),
+          display: new Map([
+            [$metadata.types.Person.methods.rename.return, CustomDisplay],
+          ]),
+        },
+      },
+    );
+
+    expect(wrapper.find(".custom-method-input").exists()).toBeTruthy();
+    vm.rename.args.name = "A";
+    await vm.rename.invokeWithArgs();
+    await nextTick();
+    expect(wrapper.find(".custom-method-display").exists()).toBeTruthy();
   });
 });